Fix RHS Thread view's textbox activation area size.
This whole bug is a by-product of some code trying to resize the formatting bar based on the width of text rows in the textbox. (Why? O.o)
The fix is to simply remove the code that tries to do that resizing.

The narrow RHS Thread view is always so narrow the formatting bar simply has no space to be shown completely,
thus is automatically reduced to show the three-dot-icon anyway.
In the wide RHS Thread view, formatting bar usually has space to be shown completely and thus is,
until the browser window is resized to be so narrow it is automatically switched to show the three-dot-icon.

So removing this piece of code doesn't seem break any correct behaviour, only fixes this faulty behaviour.